+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildImageMojo.java
@@ -32,7 +32,6 @@ import com.google.cloud.tools.jib.plugins.common.InvalidContainerVolumeException
 import com.google.cloud.tools.jib.plugins.common.InvalidWorkingDirectoryException;
 import com.google.cloud.tools.jib.plugins.common.MainClassInferenceException;
 import com.google.cloud.tools.jib.plugins.common.PluginConfigurationProcessor;
-import com.google.cloud.tools.jib.plugins.common.RawConfiguration;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Strings;
 import java.io.IOException;
@@ -97,7 +96,6 @@ public class BuildImageMojo extends JibPluginConfiguration {
 
     try {
       AbsoluteUnixPath appRoot = MojoCommon.getAppRootChecked(this);
-
       MavenProjectProperties projectProperties =
           MavenProjectProperties.getForProject(
               getProject(),
@@ -107,11 +105,13 @@ public class BuildImageMojo extends JibPluginConfiguration {
               appRoot);
       EventDispatcher eventDispatcher =
           new DefaultEventDispatcher(projectProperties.getEventHandlers());
-      RawConfiguration rawConfiguration = new MavenRawConfiguration(this, eventDispatcher);
 
       PluginConfigurationProcessor pluginConfigurationProcessor =
           PluginConfigurationProcessor.processCommonConfigurationForRegistryImage(
-              rawConfiguration, projectProperties);
+              new MavenRawConfiguration(this),
+              new MavenSettingsServerCredentials(
+                  getSession().getSettings(), getSettingsDecrypter(), eventDispatcher),
+              projectProperties);
 
       ImageReference targetImageReference = pluginConfigurationProcessor.getTargetImageReference();
       HelpfulSuggestions helpfulSuggestions =