+++ jib-maven-plugin/CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - `gwt-app` packaging type now builds a WAR container.
+- When building to Docker and no `<to><image>` is defined, artifact ID is used as an image reference instead of project name.
 - `<useOnlyProjectCache>` parameter is deprecated in favor of the `jib.useOnlyProjectCache` system property ([#1308](https://github.com/GoogleContainerTools/jib/issues/1308))
 
 ### Fixed